pyproject.toml
[project]
name = "jira-daily-task-extractor"
version = "0.1.0"
description = ""
authors = [
    {name = "Ash_K",email = "77124115+tsg-ash-kanagat@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "certifi (==2024.7.4)",
    "charset-normalizer (==3.3.2)",
    "defusedxml (==0.7.1)",
    "idna (==3.7)",
    "jira (==3.8.0)",
    "numpy (==2.0.1)",
    "oauthlib (==3.2.2)",
    "packaging (==24.1)",
    "pandas (==2.2.2)",
    "pillow (==10.4.0)",
    "python-dateutil (==2.9.0.post0)",
    "python-dotenv (==1.0.1)",
    "pytz (==2024.1)",
    "requests (==2.32.3)",
    "requests-oauthlib (==2.0.0)",
    "requests-toolbelt (==1.0.0)",
    "six (==1.16.0)",
    "typing-extensions (==4.12.2)",
    "tzdata (==2024.1)",
    "tzlocal (==5.2)",
    "urllib3 (==2.2.2)",
    "tabulate (>=0.9.0,<0.10.0)",
    "streamlit (>=1.41.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
